FRONT-8: styling add page dialog
