+++ src/editor/dialogs/add-page-dialog/components/add-page-dialog.scss
@@ -1,0 +1,2 @@
+.add-page-dialog {
+}
