+++ src/editor/dialogs/add-page-dialog/state/AddPageDialogState.ts
@@ -1,0 +1,9 @@
+import { action, observable } from 'mobx';
+
+export class AddPageDialogState {
+  @observable public pageName = '';
+
+  @action public setName(name: string) {
+    this.pageName = name;
+  }
+}
