+++ src/editor/editor-root/components/EditorRoot.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { Button, MenuItem, NonIdealState } from '@blueprintjs/core';
 import { observer } from 'mobx-react';
 
-import { AddPageDialog } from '../../dialogs/add-page-dialog/AddPageDialog';
+import { AddPageDialog } from '../../dialogs/add-page-dialog/components/AddPageDialog';
 import { DockableUI } from '../../dockable-ui/components/DockableUI';
 import { EditorNavbar } from './EditorNavbar';
 import { EditorRootState } from '../state/EditorRootState';
